requirements.txt
arrow==1.3.0
asgiref==3.9.1
binaryornot==0.4.4
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==45.0.6
Django==5.2.4
django-allauth==65.10.0
django-browser-reload==1.18.0
django-tailwind==4.2.0
django-widget-tweaks==1.5.0
psycopg[binary]>=3.1,<4
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pillow==11.3.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250708
tzdata==2025.2
urllib3==2.5.0
gunicorn
whitenoise
python-dotenv
django-crispy-forms
crispy-tailwind
